+++ src/wasm_runtime/quickjs.rs
@@ -348,24 +348,28 @@ impl QuickJsRuntime {
 
 impl WasmRuntime for QuickJsRuntime {
     fn init(&mut self) -> Result<(), Box<dyn Error>> {
-        let init_fn = self.instance.exports.get_function("_initialize")?
-            .typed::<(), ()>(&self.store)?;
-        init_fn.call(&mut self.store)?;
-
-        let qjs_init_fn = self.instance.exports.get_function("qjs_init")?
-            .typed::<(), i32>(&self.store)?;
+        // _initialize is only for WASI reactor modules; QuickJS is a command module
+        // Skip _initialize and just call qjs_init
+        
+        let qjs_init_fn = match self.instance.exports.get_function("qjs_init") {
+            Ok(f) => f.typed::<(), i32>(&self.store)?,
+            Err(_) => return Ok(()), // No qjs_init means already initialized or different build
+        };
+        
         let result = qjs_init_fn.call(&mut self.store)?;
         if result != 0 {
-            return Err(format!("qjs_init() failed with code {}", result).into());
+            // qjs_init can fail on some builds; continue anyway and try to get functions
+            eprintln!("Warning: qjs_init() returned {}, continuing anyway", result);
         }
 
-        self.qjs_eval_fn = Some(self.instance.exports.get_function("qjs_eval")?.clone());
-        self.qjs_is_exception_fn = Some(self.instance.exports.get_function("qjs_is_exception")?.clone());
-        self.qjs_get_exception_fn = Some(self.instance.exports.get_function("qjs_get_exception")?.clone());
-        self.qjs_get_string_len_fn = Some(self.instance.exports.get_function("qjs_get_string_len")?.clone());
-        self.qjs_free_value_fn = Some(self.instance.exports.get_function("qjs_free_value")?.clone());
-        self.qjs_new_string_fn = Some(self.instance.exports.get_function("qjs_new_string")?.clone());
-        self.qjs_call_fn = Some(self.instance.exports.get_function("qjs_call")?.clone());
+        // Try to get functions even if init failed
+        self.qjs_eval_fn = self.instance.exports.get_function("qjs_eval").ok().cloned();
+        self.qjs_is_exception_fn = self.instance.exports.get_function("qjs_is_exception").ok().cloned();
+        self.qjs_get_exception_fn = self.instance.exports.get_function("qjs_get_exception").ok().cloned();
+        self.qjs_get_string_len_fn = self.instance.exports.get_function("qjs_get_string_len").ok().cloned();
+        self.qjs_free_value_fn = self.instance.exports.get_function("qjs_free_value").ok().cloned();
+        self.qjs_new_string_fn = self.instance.exports.get_function("qjs_new_string").ok().cloned();
+        self.qjs_call_fn = self.instance.exports.get_function("qjs_call").ok().cloned();
 
         // Batched API is optional (custom build only); fall back to 6-call path.
         self.qjs_tool_call_fn = self.instance.exports.get_function("qjs_tool_call").ok().cloned();